backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
PyMySQL
SQLAlchemy
httpx
pydantic
python-multipart
cors
fastapi-cors